requirements.txt
flask
langchain-groq
langgraph
langchain-core
python-dotenv
langchain-community
PyPDF2
